% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_select.R
\name{f_model_select}
\alias{f_model_select}
\title{R package demo for Forward Stepwise Selection}
\usage{
f_model_select(data)
}
\description{
R package demo for Forward Stepwise Selection
}
\author{
Chen Zhou
}
